Hoang anh

+++ AIMS/Backend/src/main/java/com/aims/service/DeliveryInfoService.java
@@ -1,11 +1,11 @@
 package com.aims.service;
 
-import com.aims.entity.DeliveryInfo.DeliveryInfo;
+import com.aims.entity.delivery.DeliveryInfo;
 import org.springframework.stereotype.Service;
 
 @Service
 public interface DeliveryInfoService {
     DeliveryInfo create(DeliveryInfo deliveryInfo);
     DeliveryInfo getDeliveryInfo(String id);
-    Double calculateShippingFee(String province, boolean isRushDelivery);
+    Integer calculateShippingFee(String province, boolean isRushDelivery);
 }
